requirements.txt
torch-summary
torch
imageio
torchvision
Ninja
numpy
scikit-learn
pandas
tqdm
pillow
opencv-python
matplotlib
gdown
nvidia-pyindex
nvidia-cuda-runtime-cu11
pyncclient